remove CommInterface and some statics
